파일 단위로 분리
